create UseManualSearch hook
